Fix compilation warnings

+++ source/cgame/cg_ents.cpp
@@ -187,9 +187,9 @@ static void CG_NewPacketEntityState( entity_state_t *state )
 	else
 	{
 		// if it moved too much force the teleported bit
-		if(  abs( cent->current.origin[0] - state->origin[0] ) > 512
-			|| abs( cent->current.origin[1] - state->origin[1] ) > 512
-			|| abs( cent->current.origin[2] - state->origin[2] ) > 512 )
+		if(  abs( (int)(cent->current.origin[0] - state->origin[0]) ) > 512
+			|| abs( (int)(cent->current.origin[1] - state->origin[1]) ) > 512
+			|| abs( (int)(cent->current.origin[2] - state->origin[2]) ) > 512 )
 		{
 			cent->serverFrame = -99;
 		}
